+++ pages/api/shortcut.ts
@@ -19,12 +19,38 @@ export default async function handler(req, res) {
 
     const shortcutName = await page.evaluate(() => document.querySelectorAll('h3')[0].textContent) as string;
     const description = await page.evaluate(() => document.querySelectorAll('h4')[0].textContent) as string;
+    const hearts = await page.evaluate(() => document.querySelector('.heart-count')?.textContent) as string;
+    const downloads = await page.evaluate(() => document.querySelectorAll('.information > p')[4]?.textContent).then((e)=>e.replace('Downloads: ', '')) as string;
+    const author = await page.evaluate(() => document.querySelectorAll('.information > p')[0]?.textContent).then((e)=>e.replace('Author:\n@', '').replace('\n', '')) as string;
+    const authorURL = await page.evaluate(() => document.querySelectorAll('.information > p')[0].querySelector('a')?.href) as string;
+    const latest_version = await page.evaluate(() => document.querySelectorAll('.information > p')[1]?.textContent).then((e)=>e.replace('Version: ', '')) as string;
+    const updated = await page.evaluate(() => document.querySelectorAll('.information > p')[3]?.textContent).then((e)=>e.replace('Updated: ', '')) as string;
+    const iOS_version = await page.evaluate(() => document.querySelectorAll('.information > p')[2]?.textContent).then((e)=>e.replace('iOS: ', '')) as string;
+    const categories = await page.evaluate(()=>
+        Array.from(document.querySelectorAll('.information > ul')[0].querySelectorAll('li'),(e)=>(e.innerText))
+    ) as string[];
+    const download_link = await page.evaluate(() => document.querySelector('.actions')?.querySelector('a')?.href) as string;
+
 
 
     await browser.close()
     res.status(200).json({
         id: id,
         name: shortcutName,
-        description: description
+        description: description,
+        hearts: Number(hearts),
+        downloads: Number(downloads),
+        author: {
+            username: author,
+            page_link: authorURL,
+            api_link: `https://rh-api.alombi.xyz/author?username=${author}`
+        },
+        latest_version: {
+            version: latest_version,
+            updated: updated
+        },
+        iOS: iOS_version,
+        categories: categories,
+        download_link: download_link
     })
 }