+++ CHANGES.md
@@ -1,3 +1,7 @@
 - complete rewrite in Typescript
 - new `download_link` parameter in `/changelog`
 - new `routinehub_api_link` parameter in `/search`
+- removes `icon` and `related` from `/shortcut`
+- `author` in `shortcut` is now an object containing `username`, `page_link`  and `author_link`
+- added `latest_version` (object formed by `version` and `updated`), `iOS` and `categories` (array) in `/shortcut`
+- added `download_link` in `/shortcut`